backend/src/requirements.txt

Django==2.2.4
django-cors-headers==3.0.2
djangorestframework==3.10.2
mysqlclient==1.4.2.post1
pytz==2019.2
sqlparse==0.3.0
django-rest-auth==0.9.5
django-allauth==0.40.0
Pillow==6.1.0
django-cleanup==4.0.0
python-decouple==3.1